Fix: Fixes in create dataset
Bugs have been fixed that caused the author to be displayed multiple times in the dataset after it was created and sent to the staging area. Additionally, the method for updating authors has been corrected when updating the dataset in the staging area or when uploading it to Zenodo from the staging area.